9p: Add a "veryloose" cache mode
This is intended to speed up NixOS VM tests, where the Nix store on the host is
mounted via 9P in the QEMU guests. Since the host's Nix store is immutable (that
is, paths in /nix/stiore cannot change, and cannot be deleted because they're
reachable from garbage collector roots on the host), we don't have to worry
about cache coherence. In particular, this allows us to:

* Cache negative dentries. This gives a huge speedup in many scenarios because
  Nix/NixOS uses large search paths in many contexts. For example, whereas in a
  typical Linux distribution, resolving "#include <foo.h>" will only hit a
  couple of directories, in Nix there can be many dozens, one for each
  dependency. Thus it's important to cache negative lookup results.

* Don't revalidate dentries on lookup. This is safe due to immutability. This
  also reduces the number of "walk" server calls a lot.

* Cache symlink reads. NixOS relies heavily on symlink trees, so caching them
  reduces the number of "readlink" server calls a lot.

Some speedups (on Nixpkgs f15d037cd24dfcc8a0958ebe8634484794ddaa81):

* nix-build nixos/tests/kde5.nix -A test

  Before: 275.71s
  After:   91.43s

* nix-build pkgs/build-support/vm/test.nix -A buildPanInVM

  Before: 219.70s
  After:  121.01s
  Native: 106.83s (i.e. "nix-build -A pan")

  Here, the number of 9P calls to the host was reduced from ~1157K to ~275K. In
  particular, the number of walk calls was reduced from ~974K to ~89K, and
  readlink calls from ~27K to 52.

Note that "veryloose" is not restricted to immutable filesystems; it's also
appropriate for any filesystem where the client is the single writer.

More improvements are possible - for example, we might be able to delay "open"
server calls until we actually need to do an uncached read, thereby reducing
traffic when reading a cached file.

(cherry picked from commit 7e20254412c780a2102761fee92cb1d32ceeaefd)