9p: Add a "forceuid" mount option
This forces the uid/gid of all files to be equal to the values specified by the
dfltuid and dfltgid mount options. This is useful in NixOS VM tests, where we
want /nix/store in the guest to appear owned by root, even though it has a
different uid/gid in the host filesystem.